Highscores, new weapon
Added highscores which are now displayed on the game over screen. New pulse laser weapon which can be charged up requiring no ammunition but very heat intensive. When weapon heat bar reaches maximum capacity the player has to wait until the heat level reaches zero before any weapons can be fired again.

+++ TopDownShooterProject/Assets/Scripts/Player.cs
@@ -10,7 +10,7 @@ public class Player : MonoBehaviour {
     public delegate void UpdateAmmoCount(int newAmmoCount);
     public static event UpdateAmmoCount OnUpdateAmmoCount;
 
-    public delegate void UpdateWeaponHeat(float newWeaponHeat);
+    public delegate void UpdateWeaponHeat(int newWeaponHeat);
     public static event UpdateWeaponHeat OnUpdateWeaponHeat;
 
     private Animator gunAnim;
@@ -51,7 +51,7 @@ public class Player : MonoBehaviour {
         }
     }
 
-    public void SendWeaponHeatData(float weaponHeat)
+    public void SendWeaponHeatData(int weaponHeat)
     {
         if (OnUpdateWeaponHeat != null)
         {
